refactor(oo): changed project groupId and contextual folders tree
