+++ oo/uninadelivery/pom.xml
@@ -4,7 +4,7 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.unina._2324oobd._22</groupId>
+  <groupId>com.unina.2324oobd.22</groupId>
   <artifactId>uninadelivery</artifactId>
   <version>0.0.1-SNAPSHOT</version>
