--- oo/uninadelivery/src/main/java/com/unina/_2324oobd/_22/Area.java
@@ -1,3 +1,0 @@
-package com.unina._2324oobd._22;
-
-public class Area { }
