+++ oo/uninadelivery/src/main/java/com/unina/oobd2324gr22/Address.java
@@ -1,3 +1,3 @@
-package com.unina._2324oobd._22;
+package com.unina.oobd2324gr22;
 
 public class Address extends Area { }
