+++ oo/uninadelivery/src/main/java/com/unina/oobd2324gr22/App.java
@@ -1,4 +1,4 @@
-package com.unina._2324oobd._22;
+package com.unina.oobd2324gr22;
 
 /**
  * Hello world!
