+++ oo/uninadelivery/src/main/java/com/unina/oobd2324gr22/Area.java
@@ -1,0 +1,3 @@
+package com.unina.oobd2324gr22;
+
+public class Area { }
