+++ oo/uninadelivery/src/main/java/com/unina/oobd2324gr22/package-info.java
@@ -3,4 +3,4 @@
  * This package contains classes related to the UniNa Delivery project.
  * It provides functionalities for managing deliveries and tracking packages.
  */
-package com.unina._2324oobd._22;
+package com.unina.oobd2324gr22;
